requirements.txt
ollama
chromadb
streamlit 
pdfplumber 
langchain
langchain-core
langchain-community
langchain_text_splitters
unstructured
unstructured[all-docs]
onnx==1.16.1
virtualenv
camelot-py
PyPDF2==2.10.0
PyMuPDF
reportlab
nltk==3.9.1
pytesseract
pillow